feat : features/template 1차 작업 완료
