+++ client/src/features/template/TemplateDetail/TemplateDetail.tsx
