+++ client/src/features/template/hooks/useTemplateManager.ts
