+++ client/src/features/template/hooks/useTemplateStoreManager.ts
