+++ client/src/features/template/index.ts
@@ -1,0 +1,1 @@
+export { default as TemplateList } from './TemplateList';
