+++ client/src/pages/coverletter/new/index.tsx
@@ -2,7 +2,7 @@ import { useEffect } from 'react';
 
 import APIErrorBoundary from '@/components/APIErrorBoundary';
 import CoverLetterCreator from '@/src/features/coverletter/new/CoverLetterCreator';
-import useTemplateStoreManager from '@/components/Admin/Template/hooks/useTemplateStoreManager';
+import useTemplateStoreManager from '@/src/features/template/hooks/useTemplateStoreManager';
 
 function CoverLetterEditorPage() {
   const { clearCurTemplate } = useTemplateStoreManager();
